Made libakonadi QT_NO_ASCII_CAST clean and changed the uid of DataReference to be an uint instead of a QByteArray
svn path=/trunk/KDE/kdepim/akonadi/; revision=589204

+++ clients/akonadi/deletecommand.cpp
@@ -30,7 +30,7 @@ DeleteCommand::DeleteCommand(const QString & uid) :
 
 void DeleteCommand::exec()
 {
-  DataReference ref( mUid, QString() );
+  DataReference ref( mUid.toUInt(), QString() );
   ItemDeleteJob delJob( ref );
   if ( !delJob.exec() ) {
     err() << "Error deleting item '" << mUid << "': "
